requirements.txt
# requirements.txt
# Зависимости проекта

# Библиотека для создания Telegram-ботов (версия 3.x)
aiogram==3.13.1

# Для работы с переменными окружения
python-dotenv==1.0.0

